requirements.txt
click
pygments
python-magic